OmnicProject

+++ src/main/java/br/com/omnic2/pedido/PedidoController.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
@@ -14,6 +15,7 @@ import org.springframework.web.util.UriComponentsBuilder;
 import javax.transaction.Transactional;
 import javax.validation.Valid;
 import java.net.URI;
+import java.util.Objects;
 import java.util.Optional;
 
 @RestController
@@ -32,6 +34,7 @@ public class PedidoController {
         return new PedidoDto().convert(pedidos);
     }*/
 
+    // ROLE_CLIENTE
     @GetMapping
     public Page<PedidoDto> readByUser(@PageableDefault(page = 0, size = 10) Pageable pageable) {
         Usuario usuario = (Usuario) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
@@ -39,6 +42,7 @@ public class PedidoController {
         return new PedidoDto().convert(pedidos);
     }
 
+    // ROLE_ADMINISTRADOR
     @GetMapping("/{id}")
     public ResponseEntity<PedidoDto> readById(@PathVariable("id") Long id) {
         Optional<Pedido> pedidoOptional = pedidoRepo.findById(id);
@@ -48,6 +52,7 @@ public class PedidoController {
         return ResponseEntity.notFound().build();
     }
 
+    // ROLE_CLIENTE
     @PostMapping
     @Transactional
     public ResponseEntity<PedidoDto> create(@RequestBody @Valid PedidoForm pedidoForm, UriComponentsBuilder uriBuilder) {
@@ -62,24 +67,34 @@ public class PedidoController {
         return ResponseEntity.notFound().build();
     }
 
+    // ROLE_CLIENTE
     @PutMapping("/{id}")
     @Transactional
     public ResponseEntity<PedidoDto> update(@PathVariable("id") Long id, @RequestBody @Valid PedidoUpdateForm pedidoUpdateForm) {
+        Usuario usuario = (Usuario) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         Optional<Pedido> pedidoOptional = pedidoRepo.findById(id);
         if(pedidoOptional.isPresent()) {
-            Pedido pedido = pedidoUpdateForm.update(pedidoOptional.get(), produtoRepo);
-            return ResponseEntity.ok(new PedidoDto(pedido));
+            if(Objects.equals(pedidoOptional.get().getUsuario().getId(), usuario.getId())) {
+                Pedido pedido = pedidoUpdateForm.update(pedidoOptional.get(), produtoRepo);
+                return ResponseEntity.ok(new PedidoDto(pedido));
+            }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
         return ResponseEntity.notFound().build();
     }
 
+    // ROLE_CLIENTE
     @DeleteMapping("/{id}")
     @Transactional
     public ResponseEntity<?> delete(@PathVariable("id") Long id) {
+        Usuario usuario = (Usuario) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         Optional<Pedido> pedidoOptional = pedidoRepo.findById(id);
         if(pedidoOptional.isPresent()) {
-            pedidoRepo.deleteById(id);
-            return ResponseEntity.ok().build();
+            if(Objects.equals(pedidoOptional.get().getUsuario().getId(), usuario.getId())) {
+                pedidoRepo.deleteById(id);
+                return ResponseEntity.ok().build();
+            }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
         return ResponseEntity.notFound().build();
     }
